feat: /status slash command + move codex quota to analytics
Two follow-ups based on user feedback to PR #823:

1. /status slash command — typing /status in the chat input prints the
   current Codex subscription quota (one line per openai-codex tier with
   primary/weekly used_percent and reset countdown) in the feedback bar.
   Reuses existing showFeedback transient pattern (matches /cwd) with 10s
   duration so the multi-line summary is readable.

2. Codex Quota card relocated — Ops is for operational actions (approvals,
   audit). The quota panel belongs with the other usage/spend metrics on
   Analytics. Card now sits between the summary grid and the token-spend
   chart, keeping 60s polling, SSE refresh, and warn/critical tints.

The new lib/codexStatus.ts formatter inlines the small reset-countdown
helper rather than importing it across files — keeps the module
standalone-loadable under `node --experimental-strip-types --test`, which
can't resolve extensionless ESM imports between source files.

Bumps VERSION → 0.32.29 + CHANGELOG.
