+++ frontend/console/src/components/Analytics.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
-  import { onMount } from 'svelte'
-  import { getAnalytics } from '../lib/api'
-  import type { AnalyticsDailyRow, AnalyticsModelRow, AnalyticsResponse, AnalyticsSkillRow } from '../lib/types'
+  import { onMount, onDestroy } from 'svelte'
+  import { getAnalytics, getCodexUsage, streamEvents } from '../lib/api'
+  import type { AnalyticsDailyRow, AnalyticsModelRow, AnalyticsResponse, AnalyticsSkillRow, CodexUsageTier } from '../lib/types'
+  import { formatResetCountdown } from '../lib/formatDuration'
   import { t } from '../i18n'
 
   const dayOptions = [7, 30, 90]
@@ -15,6 +16,11 @@
   let selectedDays = $state(7)
   let loading = $state(true)
   let error = $state('')
+  let codexTiers: CodexUsageTier[] = $state([])
+  let codexLoading = $state(false)
+  let codexError = $state('')
+  let codexTimer: ReturnType<typeof setInterval> | null = null
+  let stopCodexStream: (() => void) | null = null
 
   let daily: AnalyticsDailyRow[] = $derived.by(() => analytics?.daily ?? [])
   let models: AnalyticsModelRow[] = $derived.by(() => analytics?.models ?? [])
@@ -80,8 +86,43 @@
     return Math.max(1, (tokens / total) * totalHeight)
   }
 
+  async function refreshCodexUsage() {
+    codexLoading = true
+    try {
+      const res = await getCodexUsage()
+      codexTiers = (res.tiers ?? []).filter((tier) => (tier.provider ?? '').toLowerCase() === 'openai-codex')
+      codexError = ''
+    } catch (err) {
+      codexError = err instanceof Error ? err.message : 'failed to load codex quota'
+    } finally {
+      codexLoading = false
+    }
+  }
+
+  function bandClass(usedPercent: number | undefined): string {
+    if (usedPercent === undefined) return ''
+    if (usedPercent >= 95) return 'codex-row-critical'
+    if (usedPercent >= 90) return 'codex-row-warn'
+    return ''
+  }
+
   onMount(() => {
     void loadAnalytics()
+    void refreshCodexUsage()
+    codexTimer = setInterval(() => { void refreshCodexUsage() }, 60_000)
+    stopCodexStream = streamEvents((event) => {
+      if (event.category === 'codex_quota') {
+        void refreshCodexUsage()
+      }
+    })
+  })
+
+  onDestroy(() => {
+    stopCodexStream?.()
+    if (codexTimer !== null) {
+      clearInterval(codexTimer)
+      codexTimer = null
+    }
   })
 </script>
 
@@ -130,6 +171,60 @@
       <small>{$t.analytics.summary.daysSuffix(selectedDays)}</small>
     </div>
   </section>
+
+  {#if codexTiers.length > 0 || codexLoading || codexError}
+    <section class="card codex-quota-section">
+      <div class="card-header">
+        <span class="card-title">Codex subscription quota</span>
+        <div class="card-header-actions">
+          <button type="button" class="btn btn-ghost btn-sm" disabled={codexLoading} onclick={() => { void refreshCodexUsage() }}>
+            {codexLoading ? 'Refreshing…' : 'Refresh'}
+          </button>
+        </div>
+      </div>
+      {#if codexError}
+        <div class="error-banner">{codexError}</div>
+      {/if}
+      {#if codexTiers.length === 0 && !codexLoading && !codexError}
+        <div class="empty-state">No openai-codex tiers configured.</div>
+      {:else}
+        <div class="codex-list">
+          {#each codexTiers as tier}
+            <div class="codex-tier {bandClass(tier.snapshot?.primary?.used_percent)}">
+              <div class="codex-tier-head">
+                <span class="codex-tier-label">{tier.tier}</span>
+                <span class="codex-tier-meta">{tier.provider ?? ''}{tier.model ? ` · ${tier.model}` : ''}</span>
+              </div>
+              {#if !tier.snapshot}
+                <div class="codex-tier-empty">Awaiting first request…</div>
+              {:else}
+                <div class="codex-windows">
+                  {#if tier.snapshot.primary}
+                    <div class="codex-window {bandClass(tier.snapshot.primary.used_percent)}">
+                      <span class="codex-window-label">5h window</span>
+                      <span class="codex-window-pct">{tier.snapshot.primary.used_percent.toFixed(1)}%</span>
+                      {#if tier.snapshot.primary.reset_after_seconds !== undefined}
+                        <span class="codex-window-reset">resets in {formatResetCountdown(tier.snapshot.primary.reset_after_seconds)}</span>
+                      {/if}
+                    </div>
+                  {/if}
+                  {#if tier.snapshot.secondary}
+                    <div class="codex-window {bandClass(tier.snapshot.secondary.used_percent)}">
+                      <span class="codex-window-label">weekly</span>
+                      <span class="codex-window-pct">{tier.snapshot.secondary.used_percent.toFixed(1)}%</span>
+                      {#if tier.snapshot.secondary.reset_after_seconds !== undefined}
+                        <span class="codex-window-reset">resets in {formatResetCountdown(tier.snapshot.secondary.reset_after_seconds)}</span>
+                      {/if}
+                    </div>
+                  {/if}
+                </div>
+              {/if}
+            </div>
+          {/each}
+        </div>
+      {/if}
+    </section>
+  {/if}
 
   <section class="analytics-chart card">
     <div class="card-header">
@@ -459,5 +554,102 @@
     .analytics-tables {
       grid-template-columns: 1fr;
     }
+  }
+
+  .codex-quota-section {
+    margin-bottom: var(--space-6);
+  }
+
+  .codex-list {
+    display: flex;
+    flex-direction: column;
+    gap: var(--space-3);
+  }
+
+  .codex-tier {
+    border: 1px solid var(--border-subtle);
+    border-radius: var(--radius-md);
+    padding: var(--space-3) var(--space-4);
+    background: var(--surface);
+    transition: border-color var(--duration-fast) var(--ease-out), background var(--duration-fast) var(--ease-out);
+  }
+
+  .codex-tier.codex-row-warn {
+    border-color: var(--warning);
+    background: var(--warning-muted);
+  }
+
+  .codex-tier.codex-row-critical {
+    border-color: var(--danger, #ef4444);
+    background: rgba(239, 68, 68, 0.1);
+  }
+
+  .codex-tier-head {
+    display: flex;
+    align-items: baseline;
+    justify-content: space-between;
+    gap: var(--space-3);
+    margin-bottom: var(--space-2);
+  }
+
+  .codex-tier-label {
+    font-weight: 600;
+    text-transform: uppercase;
+    font-size: var(--text-xs);
+    letter-spacing: 0.04em;
+    color: var(--primary-text);
+  }
+
+  .codex-tier-meta {
+    color: var(--text-tertiary);
+    font-size: var(--text-xs);
+    font-family: var(--font-mono);
+  }
+
+  .codex-tier-empty {
+    color: var(--text-tertiary);
+    font-style: italic;
+    font-size: var(--text-sm);
+  }
+
+  .codex-windows {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(180px, 1fr));
+    gap: var(--space-3);
+  }
+
+  .codex-window {
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+    padding: var(--space-2) var(--space-3);
+    border-radius: var(--radius-sm);
+    background: var(--surface-subtle, rgba(255, 255, 255, 0.03));
+  }
+
+  .codex-window.codex-row-warn {
+    background: var(--warning-muted);
+  }
+
+  .codex-window.codex-row-critical {
+    background: rgba(239, 68, 68, 0.18);
+  }
+
+  .codex-window-label {
+    font-size: var(--text-xs);
+    color: var(--text-tertiary);
+    text-transform: uppercase;
+    letter-spacing: 0.03em;
+  }
+
+  .codex-window-pct {
+    font-size: var(--text-lg);
+    font-weight: 600;
+    font-family: var(--font-mono);
+  }
+
+  .codex-window-reset {
+    font-size: var(--text-xs);
+    color: var(--text-secondary);
   }
 </style>
